// src/taixu/components/mainview/types.ts
export interface RichTextSettings {
  quoteColor: string;
  singleStarColor: string;
  doubleStarColor: string;
  bracketColor: string;
  quoteBold: boolean;
  quoteItalic: boolean;
  singleStarBold: boolean;
  singleStarItalic: boolean;
  doubleStarBold: boolean;
  doubleStarItalic: boolean;
  bracketBold: boolean;
  bracketItalic: boolean;
  uiHeight?: number;
  uiWidth?: number;
  sidebarWidth?: number;
  secondaryModalWidth?: number;
  secondaryModalHeight?: number;
  topBarGap?: number;
}